Upload certificates to bluemix after retrieval.
This commit also had to reverse changing all of the
CLI calls to the `bx` cli.

In order to make the certificate upload work
properly, I also needed to change the way that
domain and host names are assembled to make the
bluemix domain name. This required special handling
for an edge case in which the first hostname is
'.', which would mistakenly be used as a subdomain.
This resulted in domains like '..example.com', which
are invalid.