Warn user if cert already in place.
Previously, if you had an SSL on the Bluemix domain already,
the script would fail to upload one and then report that it
succeeded because it could see a certificate after that (the
one that previously existed). I didn't want to automatically
remove someone's certificate, so now the script warns the
user, provides optional instructions on how to continue,
and exits.