Implement better checks for SSL existence
The script now checks with both the Bluemix CLI and the domain name
itself for the existence of an SSL before deciding that there isn't
one. This is necessary because of instability in the Bluemix CLI's
backend systems that can cause checking for the existence of a cert
to fail even when it does exist.

The script also now warns users that their certificate and private
key are in the current working directory and offers a suggestion
for how to clean them up.

Additionally, the certificate upload now also uploads the intermediate
certificate, which yeilds an "A" score from Qualys SSL Labs.